+++ src/WebJobs.Script/Workers/Http/Configuration/HttpWorkerOptions.cs
@@ -2,10 +2,13 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.Azure.WebJobs.Hosting;
 
 namespace Microsoft.Azure.WebJobs.Script.Workers.Http
 {
-    public class HttpWorkerOptions
+    public class HttpWorkerOptions : IOptionsFormatter
     {
         public CustomHandlerType Type { get; set; }
 
@@ -29,5 +32,14 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Http
         public bool EnableProxyingHttpRequest { get; set; }
 
         public TimeSpan InitializationTimeout { get; set; } = TimeSpan.FromSeconds(30);
+
+        public string Format()
+        {
+            return JsonSerializer.Serialize(this, typeof(HttpWorkerOptions), HttpWorkerOptionsJsonSerializerContext.Default);
+        }
     }
+
+    [JsonSourceGenerationOptions(WriteIndented = true)]
+    [JsonSerializable(typeof(HttpWorkerOptions))]
+    internal partial class HttpWorkerOptionsJsonSerializerContext : JsonSerializerContext;
 }